+++ src/main/java/io/github/astrapi69/entity/pairable/KeyValueOwnerEntity.java
@@ -26,8 +26,8 @@ package io.github.astrapi69.entity.pairable;
 
 import java.io.Serializable;
 
-import javax.persistence.Column;
-import javax.persistence.MappedSuperclass;
+import jakarta.persistence.Column;
+import jakarta.persistence.MappedSuperclass;
 
 import lombok.AccessLevel;
 import lombok.EqualsAndHashCode;