+++ src/main/java/io/github/astrapi69/entity/validatable/versionable/VersionableValidatableEntity.java
@@ -26,9 +26,9 @@ package io.github.astrapi69.entity.validatable.versionable;
 
 import java.io.Serializable;
 
-import javax.persistence.Entity;
-import javax.persistence.MappedSuperclass;
-import javax.persistence.Version;
+import jakarta.persistence.Entity;
+import jakarta.persistence.MappedSuperclass;
+import jakarta.persistence.Version;
 
 import lombok.EqualsAndHashCode;
 import lombok.Getter;