+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 ## Change log
 ----------------------
 
+Version 5
+-------------
+
+CHANGED:
+
+- update jdk version to 17
+- upgrade dependency javax.persistence with new jakarta.persistence in version 3.1.0
+
 Version 4.1
 -------------